+++ app/Modules/Documentation/Views/documentation/Release-Notes.md
@@ -2,6 +2,10 @@ Release Notes
 ---
 ---
 
+### BillingTrack 7.0.2
+- replace forked admin-lte-v4 with origin "admin-lte": "^4.0.0-beta1"
+- fix trashcolumndefs for expenses
+
 ### BillingTrack 7.0.1
 - upgrade to Laravel 11
 - add merchant StripeV3